[SiFive][RISCV] Add InstCombine to merge vmv.v.x/vfmv.v.f intrinsics with arithmetic intrinsics
This allows .vx instructions to be used.

Still need to support compares, vmerge/vfmerge, vmadc, vmsbc, and
masked intrinsics. Maybe vmv.vv as well.

We may want to consider a lookup table to map intrinsics to an
"intrinsic class" which I think BSC did.

This covers a large portion of TOOLCHAIN-1359