[SiFive][RISCV] Enable address type promotion in CodeGenPrepare.
This enables the address type promotion previously only used by
AArch64. As an exampe this allows (sext (add nsw X, C)) ->
to be folded to (add nsw (sext X), C) when the sext is used by
a GEP. This can provide better opportunities to fold C into a
load/store offset. This can help RV64 for code that indexes an
array with a 32-bit int.

To make this work I also had to enable folding extends with loads
otherwise codegen prepare doesn't track the instructions.

I've also made truncating i64 to i32 be considered free for RV64.
In many cases W instructions should be able to make up for any
extends that are needed to type legalize truncated operations. This
gives the promotion more freedom to when the sext input has other users.