[SiFive][RISCV] Add peephole to merge PseudoLLA with load/store.
Limited to the non-PIC case for -mcmodel=medany.

I've restricted this to the PseudoLLA only having a single use,
but we could duplicate it at the cost an additional auipc if
linker relaxation to gp relative addressing doesn't kick in.

I had to fix the definition of PseudoStore to have the store
value in ins instead of outs. I uploaded that part to phab as
D107309.