[SiFive][RISCV] Use T0/T1 physical registers for (SP - VLEN x n) in prologue. (#121)
* Use T0/T1 physical registers for (SP - VLEN x n) in prologue. (#121)

Fix the issue TOOLCHAIN-1354.

* Avoid using virtual registers in prologue/epilog. (#129)

If the offset is larger than 12-bits immediate, the offset could not be
encoded into the SP adjustment instructions. It will generate a sequence
of instructions using virtual registers to get the immediate in a
register. If we use the virtual regiser for the SP adjustment, it may
generate store and load from different addresses in the extreme case.

// store $x20, imm($sp)
// sequence of instructions to get immediate (using $x20 for virtual registers)
// addi $sp, $sp, $x6
// load $x20, imm($sp)   // $sp is different from the store instruction