[SiFive][RISCV] Teach RISCVCodeGenPrepare to optimize widened trip counts
Widened trip counts get zero extended and they are often proceeded
by a check to make sure they are positive and greater than zero
before entering the loop. We can use that to convert the zero extend
to a sign extend which is better supported by RISCV. This sign extend
will be CSEd later with the sign extend required by the compare or
optimized out with a W instruction.

If loop unrolling has occurred there is often an i64 And to mask off
some LSBs from the wide trip count. This mask will have the upper 32
bits as zero to due to the zext. Since we know the upper 32 bits are
zero, we can fill the mask in with ones if it would allow us to use
an ANDI.