[SiFive][RISCV][InstCombine] Fix InstCombine for vector shifts with splat LHS and iXLen RHS.
To support i64 vectors on RV32, we need to truncate the XLen shift amount
for vsll, vsrl and vsra.

We also need to truncate XLen shift amounts for vnsra/vnsrl when
converting from an i32 vector an i16 vector on RV64.

I've updated the tests to cover RV32/RV64 by using iXLen and using
sed to replace it with the right type.