[SiFive][RISCV] Add an ADDIW isel pattern to reverse an optimization by ComputeNumSignBits.
This select ADDIW for (add X, -1) when X is known to be positive
and has at least 33 sign bits. If that is the case we know, that
the result will also have at least 33 sign bits. So ADDI and ADDIW
should produce the same result. By using ADDIW we can make
RISCVMIPeephole able to also understand that this produces 33 sign
bits. The other option is to teach RISCVMIPeephole to find out that
a number is positive in addition to sign extended, but that seems
like more work.