[SiFive][RISCV] Add option to force tail undisturbed policy for all vsetvli.
This can be useful on CPUs that treat tail agnostic the same as
undisturbed. It can help remove vsetvli instructions that change
the tail policy.

NOTE: Intrinsics that don't have an operand for the tail value will
not have control of what register/value is used for the tail elements.
This option should only be used to reduce vsetvli transitions. It
cannot provide tail undisturbed support where the intrinsic interface
does not give a tail undisturbed option.

A similar patch had been previously proposed, but was voted down. I
disagree with that decision and I believe this option can be useful.

My reasoning is that if hardware is allowed to treat tail agnostic
as undisturbed, the compiler should have the same freedom.

An intelligent compiler could use -mtune to automatically adjust tail
policy to minimize vsetvli transitions when tuning for a cpu that
implements tail agnostic and tail undisturbed as suggested in
the discussions in TOOLCHAIN-1415. That's going to require some backtracking
in our vsetvli insertion algorithm to do well.