[SiFive][RISCV] Add a __builtin_rvv_is_splat intrinsic to allow Recode optimization
This allows Recode to specialize their implementations for when the
compiler can determine that some inputs come from a vmv.v.x or vfmv.v.f.

This is conceptually similiar to __builtin_constant_p.

This adds a new IR intrinsic riscv.is.splat. InstCombine will fold it
to true as soon as it can. InstCombine cannot fold it to false because
it may require inlining before it can be proven true. So we must allow
many optimizations to run first.

I've modified the existing lower constant intrinsics pass to fold this
intrinsic to true or false and fix up control flow. This will prevent
the intrinsic from reaching SelectionDAG and needing to emit runtime code.