[SiFive][RISCV][InstCombine] Call SimplifyDemandedBits on scalar operands of vector shift intrinsics.
Similar to what was done in SelectionDAG recently. We may have ANDs in the
code before SelectionDAG that can be optimized especially if they are in
another basic block that SelectionDAG can't see.

While there replace (1 << Log2_32(TypeWidth)) with just TypeWidth which
should be equivalent if TypeWidth is a power of 2.