[SiFive][RISCV] Don't set Fast flag for vectors from allowsMisalignedMemoryAccesses.
While the spec allows misaligned vector accesses, I'm not sure that
all alignments are the same. For example, Bullet requires an additional
access if the vector is not naturally DLEN aligned.

My main motivation for this right now is to prevent DAGCombiner from
turning scalar stores into vector stores when -riscv-vector-bits-min
is on. This will prevent the issue seen in TOOLCHAIN-1445 though I'm
still not convinced the inline assembly code is safe. I'm also not
convinced that using a vsplat of 0 and a vector store of 2 elements
was good there either.

This doesn't prevent all cases of scalar stores being merged. If
they happen to be sufficiently over aligned that the vector would
be naturally aligned, it will still get merged. But in that case
each piece would be naturally DLEN aligned so that would be the
optimal case for Bullet.

I had to modify a test that was previously testing this code path
to make sure it didn't crash on RV64. I couldn't see a way to salvage
the test. Increasing the alignment will use i64 on RV64 instead.