[SiFive][RISCV] Enable Custom type legalization for f32 fixed vector intrinsics with StdExtF not StdExtD.
There's a typo here that should have been StdExtF instead of
StdExtD.

With this typo we crash when D extension is not enabled but a
f32 vcast to/from fixed intrinsic is used by Recode.

We need to change the predicate here to check the Zve extension
instead of the scalar extension, but I'll leave that for another
patch.