[SiFive][RISCV] Implement short forward branch optimization for sifive-7-series.
This is the more general optimization where ALU ops can be predicated
rather than only c.mv.

This version supports XOR on top of c.mv. This is needed because
our coremark codegen already predicates some important xors by
accident and we'll regress performance if we lose that.

I reordered the operands on MOVCC to put the $falsev in the same
operand for CCXOR. I also renamed MOVCC to CCMOV. So that both ops
start with PseudoCC as I envision adding more.