[SiFive][RISCV] Add ReadIALU to COPY in RISCVSched*.td.
This needed so that it picks up the ReadAdvance. Otherwise it is
modeled as seeing the full latency of the predecessor instruction
instead of just the ALU latency.

We should check the register classes of the COPY and
pick different classes, but I don't see any other target doing that.
Only a FIXME in AArch64's Falkor scheduler model.