[SiFive][RISCV] Add RISCVProcFamilyEnum and add all SiFive families.
This is a cherry-pick of D117060 extended for SiFive CPUs that aren't
upstream. I've replaced the existing X280 Proc feature with this.

Unclear if we need a separate X280 family or if we could just use
SiFive7 and rely on vector instructions not showing up on non-vector
CPUs. Added a FIXME for that. P270 should also be considered a vector
family.