[SiFive][RISCV][LoopPeel] Initial epilog peeling implementation.
Epilog peeling is implemented under loop unrolling and can be
generally utilized for a number of loop transformations as
it removes control flow from the main execution of loops.
The support path is highly similar to that of prolog peeling
which is the peeling algorithm currently enabled for all
targets unless specifically disabled by pragma or command line
options. The implmentation is currently only enabled for RISCV
by the switch -allow-epilog-peeling=true.