[SiFive] Remove RISCVISD::GlobalBaseReg. Directly lower to getRegister.
RISCVISelDAGToDAG.cpp can create constant pool loads for integers
during isel now. In some configurations, this produces a
RISCVISD::GlobalBaseReg in the middle of isel. If it doesn't get
topologically sorted into the isel DAG order, it won't get selected.

We could sort the node, other targets do it. But in this case it
seemed simpler to just remove the intermediate RISCVISD::GlobalBaseReg
opcode and directly lower it to a RegisterSDNode.