[SiFive][RISCV][LTO] Initial implementation of RISCV AoS to SoA framework
This change introduces the initial framework of AoS to SoA
transformations.  This first component detects loop based accesses
that are GEP related of arrays of structs and their data members.
To qualify the data members must all be scalar typed entities with a
single exception, that is complex data types which are of a variant
of fp typed pair variables.  Complex components cannot have more than
two components and must have the same internal type for the pair.