[SiFive] Teach RISCVGatherScatterOpt to recognize strided vp.gather/scatter
This is basically a copy of the fixed vector path with the root of
the recursion modified to look for step vector instead of a strided
constant. Couple other changes to check for preheader existence instead
of loop simplify form and to not require the recurrence step to
be loop invariant. Those are both probably worthwhile changes to upstream,
that's for another day.

Code kept independent for expedience of implementation.
I'm hoping vectorizer can directly emit vp.strided.load/store and
this code can be deleted.

This can probably only handle the vector induction variable being
used directly by the GEP that feeds the vp.gather/scatter. Anything
else likely uses VP instructions for arithmetic and the recursive
matcher isn't looking through those.

I need to add caching if the GEP is used by more than one gather/scatter
so that we don't expand it twice. We're too late in the pipeline to get
any CSE to trigger after this.