[SiFive] Disable shouldSinkOperands on 8 and 9 series cores.
This function makes CodeGenPrepare to reverse LICM to allow
SelectionDAG to see more .vx instruction oppportunities.

TGV has scalar to vector transfer buffers which are a limited
resource. It is not always a good idea to aggressively reference
scalars from vector instructions on this architecture.

We didn't have shouldSinkOperands implemented in the Narwhal branch
so this restores the Narwhal behavior for these CPUs.

This was requested in the #compiler-tgv-perf slack channel.