[SiFive] Add Zfh along with Zvfh in RISCV.td.
This prevents an error if you pass -march=rv64gc -mcpu=sifive-x280

The -march suppresses -mcpu from looking up default -march, but
we still infer features when the CPU name reaches the backend.
We'll see Zvfh, but won't have a Zfh or Zfhmin to go with it.

That probably still doesn't make sense. Maybe we should change
-mcpu to -mtune if a -march is provided? Or make -march and -mcpu
together an error.