[SiFive][RISCV] Add isel patterns to select (vp.merge mask, (smax X, Y), Z, VL)) to TU vmax.vv.
This removes an instruction from VLA code for 456.hmmer.

I've added more patterns then just smax because it was the easiest
place to insert the tablegen class.

This is not a scalable way to do this.

We should probably have vp.smax/smin/umax/umin intrinsics which
would change the pattern here.