[SiFive][Merge][Unify] Support LLVM IR intrinsics for xsfvcp extension.
This patch also unifies with the downstream patch.

The sf.vc.x and sf.vc.i intrinsics don't have any vector types in the output
and inputs, but the intrinsic spec said that we still need to add vetvli for
them. Therefore, we need to encode the VTYPE to these IR names.

Differential Revision: https://reviews.llvm.org/D147934