[SiFive][RISCV][LTO] Additions to legalization, detection and finding references for RISCV AoS to SoA framework.
This change completes finding all references by augmenting phi walks to
identify container type cases where a pointer was modified by argument
promotion, and by discovering and capturing argument types by arg position
for both opaque and legacy pointers which is done to discover divergent
param type mapping.  With these changes all references are found for
both container types and for array types of structs which we are targeting
for this transformation.  We also extended legalization for loads to add
an additional cases of arguments passed as pointers.  We extended IV
identification for cases where latch cmp is not in the latch block but
still controls loop exits.  Finally we translate and map the container
struct type to be used as a replacement for both the original container
type and the array type, thus reducing the number of GEPs required
and making memory contigous for the identified fields.