[SiFive][RISCV] Copy 'and X, 0xffffffff' into using BBs if it will form shXadd.uw/add.uw/slli.uw.
With the Zba instructions we can fold zexts into some instructions.
In IR this can be either a zext from i32 to i64, or an i64 And with
0xffffffff.

This patch teaches our custom CodeGenPrepare patch to duplicate Ands
into the basic blocks of instructions that use it if we can form
one of these instructions. We already had a similar transform for zext.
This is needed to work around SelectionDAG only working on a single basic
block.

I've limited to GEPs and shift left users. Could probably extend to Add,
but I haven't spotted one in the wild yet. I think I've only seen the shift
case with Zext not And, but I kept it for consistency.