[RISCV] Add support for indexed load macrofusion for future CPU evaluation
This supports fusing add and most Zba instructions with scalar load/store.
This does not include slli.uw from Zba.

I've broken it into 2 modes: only add and add plus Zba. This is based on an
assumption that the shift/zext required for Zba is more complicated for
hardware. This split may be refined in the future based on performance
evaluation.

SCT-1467