[SiFive][RISCV] Replace use of VP_SPLICE from VP_REVERSE lowering.
And improve the efficiency of the case that used it.

This case needs to split an LMUL=8 SEW=8 into two LMUL=4 SEW=8 so we
can use LMUL=8 SEW=16 for the vrgatherei16 index vector. The split
requires special handling because the VL can be any value from 0 up
to and including the VLMax for LMUL=8 SEW=8.

To do the split we do two LMUL=4 SEW=8 reverses of the whole LMUL=4
register. Then we concatenate them back together by swapping the
high and low part of the split. At this point element 0 from the original
vector will be at element VLMax-1. Element VLMax-1 from the original
vector will be in element 0. If VL!=VLMax there are VLMax-VL elements in
the lower elements that came from the upper elements of the original
vector. These need to be discarded by shifting the whole vector down
by VLMax-VL. It's safe to do the slidedown even if VL==VLMax since the
slide amount will be 0.