[SiFive][RISCV] Fold VP_MERGE_VL with a vmnot condition.
We can swap the true/false operands to remove the vmnot. Need
to be careful because VP_MERGE_VL represents a tail undisturbed
operation and the "false" operand is also the "tail" operand.
To swap the operands we need the "true" operand to match the "tail"
operand.

Instead of creating a new version of VP_MERGE_VL with a true operand
tied to tail, I've added a version that takes merge separately and
used an ISD::FREEZE to allow the operand to be used twice. This will
probably be more useful for other optimizations we may want to do.

Fixes SCT-1449