[SiFive][RISCV] Move combineVECTOR_SHUFFLEToVnsrl to lowering and support FP.
The vector_shuffle should be visited before it's inputs so we are
able to detect the extract_subvector pattern during lowering.

Doing this in lowering means we can avoid some of the legal type
checks from the combine version. We do need to check ELEN though
which the combine version missed.

This also fixes a crash in the combine if the extract_subvector
source was an lmul=16 type. The combine only checked the result
type was legal, not the input.

It also adds an explicit check that the source is twice as wide
as the input.