[SiFive][RISCV] Restrict when RISCVGatherScatterLowering can reuse an existing scalar phi
We try to find an existing scalar PHI to reuse since we can't rely
on LSR to rewrite the scalar phis due to vsetvli being in the loop.

This feature is only valid if we don't need to change the start value
based on any add/or/shl/mul that we looked through on the way to the
phi. If we do need to change the start value we can't reuse the phi
since the update will break the existing users.

This patch attempts to restrict the reuse to cases at are safe.
Namely that we didn't look through an add/or and the start value is 0.