[DSE] Prepare DSE for changes in upstream D132700.
A pre-commit test addition was requested for the above review for which
I had to revert part of the change just for Intrinsic::masked_store in
isMaskedStoreOverwrite.  Also in that review is a request to fold the new test in
masked-dead-store-no-merge.ll into masked-dead-store.ll, that change will
arive from upstream, and so only a source change is made plus the new test
file is deleted.