[SiFive][RISCV] Add scheduling information to PseudoCCMOVGPRNoX0.
This is used for Mallard's cmov optimization.

As far as I understand, the branch portion is still sent to the
branch unit and writes a predicate result. The predicate result
is consumed by a separate "move" operation that selects either
the old value or the new value based on the predicate.

I've modeled this as 2 microops with latency 2. I gave all 4
operands the same Read class, but I guess the old/new values aren't
needed until after the branch is resolved so we could have a
separate ReadAdvance for those.