[SiFive] call getMask instead of getOperand for VPBranchOnMaskRecipe (#656)
`VPBranchOnMaskRecipe` has 0 operands if the mask is all ones
and 1 operand otherwise. `getPredicatedMask()` called `getOperand()`
on a `VPBranchOnMaskRecipe` even if the mask was all ones. This would
lead to an assertion failure by `getOperand()` in the case that the mask
was all ones. Instead, `getMask()` should be called because it returns
`nullptr` if the mask is all ones, and the actual mask otherwise.