[SiFive][RISCV] Lower fdiv latency in p650 scheduler model.
p650 uses a radix-64 divider for FP so it calculates 6-bits of
the quotient per cycle. p550 uses a radix-4 divider.

I'm not sure if the ResourceCycles should be 1 less like it was
before or not. I've made them the same for now.

Adjust latencies using the table here
https://github.com/sifive/arch-specs/blob/master/core/fpdiv_mas.adoc#14-a-note-on-latency

NOTE: p550 uses a different radix-4 divider than the one this
table describes. We should not uses this table for p550.