[SiFive][LV] Skip vectorizer in pre-link stage for loops that have only non-unit strided memory access(es)
Within SiFive, we have AOS to SOA transformation that is only
effective during LTO phase. The LoopVectorizer is executed both in
pre-link and LTO. However we don't want vectorization to potententially
scramble the code and paralyze AOS to SOA transformation. This adhoc
approach is driven by SCT-1716, we seek to skip the vectorizer when
all memory accesses are non-unit strided during the pre-link stage.

The adhoc skipping is guarded by option
"sifive-vectorize-assume-optimizable-strided-accesses", whose default
value is set to false.