[SiFive][RISCV] Add experimental alternate register allocation orders for testing.
Add an alternate order to GPR to allocate the RA register (X1)
before the callee-saved registers. Since RA will always be saved
if there are any calls in the function, adding it earlier in the
allocation allows it to be used for other things between calls.

Add an alternate order for FPR that uses the compressible argument
registers first in descending order. This increases the chance of
using compressed load/store instructions.