[SiFive][LoopVectorize] Support vectorization of loops blocked by dependence distance (#611)
Vectorizes loops that were once blocked by dependence distance. This is accomplished by calculating the maximum safe VL based on the maximum safe distance imposed by dependencies. We set the VL as the minimum of the VL in the VPlan with the maximum safe VL. Then the pointer into the array is incremented by the correct number of elements on each iteration.

Co-authored-by: Michael Maitland <michael.maitland@.sifive.com>