[SiFive][RISCV][LTO] Add amendments enabling AoS to SoA for various metrics.
This change addresses two code quality issues wrt to reuse of GEP and Base
Loads of field address containers.  The resultant code generation now closely
resembles the manual versions of AoS to SoA that mimic this optimization.
Also included here is a correction for placement of ptr Loads of Base
addresses to match the pre optimized layout of such address loads.
This repairs an issue found by fetching a pointee which is a Base
address of a field array too early, causing incorrect data usage, as
the pointee changes after the GEP is formed via realloc.

SCT-1700