[SiFive][Clang][LV][RISCV] Add interface #pragma rvv lmul_sew(LMUL, SEW)
Legal identifier of LMUL is one of 'mf8', 'mf4', 'mf2', 'm1', 'm2',
'm4', or 'm8', that corresponds to integer value of -3, -2, -1, 0,
1, 2, and 3, respectively.

Legal identifier of SEW is one of 'e8', 'e16', 'e32', or 'e64', that
corresponds to integer value of 8, 16, 32, and 64, respectively.

The valid (LMUL, SEW) pair will map to its corresponding VF and clang
will interpret as enabling #pragma vectorize_width(VF, scalable).