[SiFive][RISCV][LTO] Add a case to defer some allocation processing.
Every malloc/calloc assocation for AoS is required to have a store
to stash its associated pointer, so defer any early processing of
ptr loads so that only stores are processed for this match scenario
to produce SoA context. The context for these loads is done
subsequently during store processing.