[SiFive][LV] SCT-1941: LoopVectorizePass fail when Compiling OpenCV (#819)
Fix a bug in VPlan's liveout construction. When VPlan builds liveouts, it only supports single exit loops at this moment. The current condition it checks whether a loop is single exit is if the loop has a single exit block, a single exiting block, and the exit block has a single predecessor. The last condition is not necessary but has not been an issue until we introduced the support of multi-versioning vector loops for dynamic strides. Now when the second vector loop is being built, the exit block no longer has a single predecessor which leads to the creation of a malformed lcssa PHI.

The general fix that removes the above unnecessary condition is considered too risky. So a more confined fix is proposed here to target exactly at our vector loop multiversioning case that triggers this bug.