[SiFive][RISCV] Quick fix to recover vwadd.wv/wx formation in some cases.
The matching code for widening instructions tries to determine that
all extends will be removed if an extend is used by multiple adds,
subs, or muls.

Unfortunately, there's an ordering issue right now where we check
the check the uses before we check what widening instruction we're
targeting. Since we're targeting vwadd.wv, the first operand doesn't
need to be an extend so we don't care about its users.

The fix here suppresses the check if an operand isn't an extend.
This misses the case where the first operand is an extend that we
can't shrink due to other users, but the second operand is an extend
we can shrink so we'd still be able to use vwadd.wv/wx.