[SiFive][InstCombine] Match (select (X < 0), (sub 1<<C, (zext X)), (zext X) to abs+zext.
Where C is the bitwidth of X. For example, if X is i8, 1 << C is 256.

If X is positive this returns the zext of X. If X is negative
this returns (256 - zext(X)). When x is negative (256 - zext(X)) is
equivalent to (zext (-X)). So this is
(select (X < 0), (zext(X)), (zext(-X)).

Pulling out the zext
(zext (select (X < 0), -X, X)) which is (zext (abs X))

Alive2: https://alive2.llvm.org/ce/z/NZLwfW

Fixes SCT-2020.