[SiFive][Recode] Remove vsmul optimization in lowerSHLSAT.
Current implementation uses vsmul to do lowerSHLSAT if the shift amount
can be placed in a scalar register. However, this is not true because it
always roundoff with a fixed amount (SEW - 1). But lowerSHLSAT should
roundoff with 0.

In addition, CanHoldMulRHS can be removed. vwmul + vnclip already check
CanUseWidenAlgo. Besides, since vwmul is used, the shift amount can
always support twice SEW range. Whether ShiftIncSize is true does not
matter.