[SiFive][RISCV] Don't nest calls to IRBuilder.Create*.
It's not specified what order operands to a function are evaluated.
gcc and clang use different orders.

If two IRBuilder.Create functions are passed as arguments to another
function, they might generate different IR depending on which compiler
is used.

This patch unnests some calls to make the order deterministic.