[SiFive][LV][SCT-1991] Allow vectorization of loops ME in VLA (#861)
Some loops with multiple exits, where it's possible to compute minimum
number of iterations to execute the loop without triggering jumps to
non-postexit, can be vectorized by VLS vectorizer.
However we're using tail-folding mechanism, where such vectorization was
not supported in that mode. Having this limitation is understandable as
trip count is rounded to the next divisible by VL number and correct handling of the
jumps needs to be done. But since we're not rounding trip count, the
limitation is not needed.

The changeset switches from the "remainder loop is not allowed" to "emit
remainder loop" if loop has multiple exits and reduces the trip count of
the main loop by 1 so that last iteration is executed in the scalar
remainder loop