[SiFive][InstCombine] Don't convert (X / Y) * Z --> (X * Z) / Y if it would bring an fdiv into a loop.
This is a modified version of D87479 which I wrote 2.5 years ago,
but was not accepted. Reviewers suggested it was a phase ordering
issue and another run of LICM would fix it.

An extra run of LICM for LTO was proposed recently in D143631.

Today Nick and I found a variation of the issue in Geekbench
which LICM can't fix. It's not quite as bad. We end up turning an
fdiv outside a loop and mul inside the loop into reciprocal outside
the loop and 2 multipies inside the loop.

We found it because it also caused accuracy with fast-math to be
bad enough to fail the test. Since the transform wasn't profitable,
I propose to avoid the fast-math issue by not doing the transform.

I'd like to move forward with this and get performance and accuracy
instead of debating the purity of InstCombine.

Fixes SCT-2225.