[SiFive][LV] Single block uncountable loop vectorization
This is the first part of a series of changes to enable auto vectorization of uncountable loops on RVV hardware. This step has the feature off by default and only targets single-block load-only loops with one pointer live out.

1) Targeted scenarios

This step only targets at strlen-like uncountable loops i.e. single-block uncountable loops with restrictions on loops being load only and having up to one pointer live out.

2) Countable loop pipeline vs. uncountable loop pipeline

The entry point of vectorizing uncountable loops is placed between canVectorizeInstrs and canVectorizeMemory so uncountable loops can share analysis up till canVectorizeInstrs on instruction safety and if-conversion safety.

For memory safety, LAA has a dependency on loops being countable. Therefore memory safety analysis for uncountable loops is routed to its own with instruction safety analysis missing from the countable one embedded. Uncountable loop memory safety is based on aliasing for this and a few upcoming changes which differs from the dependence testing framework for countable loops. There is some subtle implication from this approach on vector code generation due to a monotonic iteration space is no longer required, which is why induction variables are limited to pointers at this moment. Runtime disambiguation is not required by this change. 

3) Components new or diverged significantly from the countable loop path

3.1) Speculation analysis

Speculation safety is unique to uncountable loops in comparison with countable loops. This stems from the nature that some scalar iterations have to execute speculatively with vectorization. For this change, only speculative loads are supported. Stores and other unsafe instructions will be supported in the future when de-vect-speculation is introduced.

3.2) Uniformity analysis

The uniformity analysis is rewritten for uncountable loops. The way how this analysis is done for countable loops is not general enough to cover uncountable loops. The major difference is the countable loop uniformity analysis grows from a root set of uniforms while the uncountable one grows from non-uniforms. Necessary live ins and live outs are also augmented to complete the uniformity analysis.

3.3) Dependence safety

Please refer to the "loop pipeline vs. uncountable loop pipeline" section above for its discussion.

3.4) Canonical IV recipe

The canonical IV recipe is integral to countable loops and built in every single vectorized countable loop. There are many places that further depend on this fact in the countable loop path such as computing IV live outs. But for uncountable loops, a canonical IV is not available due to the fact that uncountable loops don't have a trip count that can be computed statically.

3.5) Trip count

This is another integral property of countable loops with many places depending on its availability but not computable for uncountable loops.

3.6) Live out

For countable loops, IV live outs are optimized based on the fact the loop is countable. For example, the vectorizer precomputes the exit value of each IV up front in the vector loop's pre-header to facilitate the computation of it final value. This convenience is no longer available for uncountable loops so live outs become built on top of their actual vector loop exiting values.

3.7) vsetvlimax

vsetvlimax is added to each vector loop iteration. The intent is to ensure VL is restored to VLMAX if first fault loads shrink VL. At this moment, vsetvlimax is hoisted out of the loop by LICM but without implications on correctness.

3.8) Cost model

Cost model is not available at this step but will be introduced in the future.