[RFC][SiFive][LV] Introduced VPlan-based cost model (#1074)
* Use VPlan-based cost model by default

Unlike current `LoopVectorizationCostModel`, which first records information in internal data structures and works on the original `Loop`, `BasicBlocks` and `Instructions`. The VPlan-based cost model
traverses `VPlan` and estimates each `VPRecipe` within. The goal of the VPlan-based cost model is to unblock heuristics tuning for more advanced optimizations on a `VPlan`, like peeling, hoisting etc.
Current implementation is as close to the `LoopVectorizationCostModel` as possible to avoid regressions.
Also the new cost model currently cannot be used to estimate the cost of the original loop.

To enable old cost model back `sifive-loop-vectorizer-use-vplan-based-cost-model=false` can be used.

* Replaced `LMULType` with `RVVPair` for future work

Refactors `LMULType` to `RVVPair` for the future commits to add VPlan-based cost
model that is going to use this struct to compute the cost of the VPlan.
Changed remark message for LMUL to be compatible with `lmul_sew` clause