[RISCV][SiFive] Support multiple version of vector crypto
Currently this patch is for vector crypto, however it
is able to be used in other extensions as well.

If the extension version is not default, the version info, e.g. 0p1,
would be encoded in its attribute string, e.g. zvkb0p1. This makes
the functions in RISCVISAInfo.cpp able to recognize different version.