[SiFive][RISCV][Driver] Warn when experimental extension version is not explicitly specified by the user.
The upstream clang/LLVM requires that experimental extensions always come with
an explicit version. However, our downstream clang/LLVM dropped that check
because some internal build scripts did not provide the correct information for
it (see SCT-401).

Recently, we need to support extensions with multiple versions, and some of
these versions are incompatible with each other. Therefore, we require users
to explicitly specify the version. However, to ensure a smoother migration, we
currently emit a warning instead of an error. In the next release, we plan to
change this behavior and emit an error instead.

Fix SCT-2482.