[SIFIVE][VP] Add vp.powi and a pass for expanding vp.powi before DAG. (#1118)
The patch is cherry-picked by D143578.

The patch uses different expanding way for vp.powi from the method of powi.
Vector powi is unrolled to multiple powi() libary calls in SelectionDAG, but the
method is not work for scalable vectors.
To support scalable vectors, the patch expands vp.powi at IR level. The
expanding way of vp.powi is based on compiler-rt/__powidf2.