[SiFive][RISCV] Remove InstCombine optimization for riscv_vfmv_v_f with with constant operand.
This optimization had been reported as an issue from Nick's team a while
ago and they are explicitly disabling all of the RISC-V because of it.

As far as I know Recode is no longer using this and the same transform
has been implemented in SelectionDAG.

Remove this transform so the algorithm team doesn't need to disable it.
I'll look at adding a disable for the SelectionDAG change in a separate patch.