[SiFive][LoopDataLayout] Handle non deterministic behavior for large transformations
In cases where much of function is rewriten, ensure that we produce the same order
of output regardless of the number of times that function is produced.

Additional changes were not made to the LocalCandidateMap as it requires two keys
for matching and not one, ergo leaving it in its current form is the best choice.

The test case provided is a reduced isolated version of quantum_hadamard, where a
large sample of input is required to show deterministic behavior in the modified code.